@@ -42,7 +42,7 @@
     "typescript": "^3.2.2"
   },
   "engines": {
-    "node": ">=8.0.0"
+    "node": ">=10.15.0"
   },
   "files": [
     ".oclif.manifest.json",
