@@ -1,9 +1,5 @@
-import { SfdxCommand, core } from '@salesforce/command';
-import util = require('util');
-import child_process = require('child_process');
+import { SfdxCommand } from '@salesforce/command';
 import ucc = require('../../../../shared/unzipConvertClean');
-
-const exec = util.promisify(child_process.exec);
 
 const tmpDir = 'mdapiout';
 
